+++ Projects/Application/Ledgers/Commands/DeleteLedger/IDeleteLedgerCommand.cs
@@ -2,6 +2,6 @@
 {
     public interface IDeleteLedgerCommand
     {
-        Task Execute(int id);
+        Task Execute(long id);
     }
 }
