+++ Projects/Application/Ledgers/Commands/UpdateLedger/UpdateLedgerCommand.cs
@@ -10,7 +10,7 @@ namespace NUCA.Projects.Application.Ledgers.Commands.UpdateLedger
         {
             _ledgerRepository = ledgerRepository;
         }
-        public async Task<Ledger> Execute(int id, LedgerModel model)
+        public async Task<Ledger> Execute(long id, LedgerModel model)
         {
             Ledger? ledger = await _ledgerRepository.Get(id) ?? throw new InvalidOperationException();
             ledger.Update(model.Name, model.Index);
